channel: unify txn watch logic
