+++ sgtypes/src/pending_txn.rs
@@ -1,78 +1,81 @@
 use crate::channel_transaction::ChannelTransactionProposal;
 use crate::channel_transaction_sigs::ChannelTransactionSigs;
+use crate::pending_txn::ProposalLifecycle::{Agreed, Applying, Negotiating};
 use libra_types::account_address::AccountAddress;
 use libra_types::transaction::TransactionOutput;
 use serde::{Deserialize, Serialize};
 use std::collections::{BTreeMap, BTreeSet};
 
-/// Every Transaction Proposal should wait for others' signature,
-/// TODO: should handle `agree or disagree`
+/// every proposal should have these states:
+/// 1. Negotiating
+/// 2. Agreed
+/// 3. Fulfilling
+/// 4. Fulfilled
 #[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub enum PendingTransaction {
-    WaitForSig {
-        proposal: ChannelTransactionProposal,
-        output: TransactionOutput,
-        // TODO: or call it vote?
-        signatures: BTreeMap<AccountAddress, ChannelTransactionSigs>,
-    },
-
-    WaitForApply {
-        proposal: ChannelTransactionProposal,
-        output: TransactionOutput,
-        // TODO: or call it vote?
-        signatures: BTreeMap<AccountAddress, ChannelTransactionSigs>,
-    },
+pub struct PendingTransaction {
+    proposal: ChannelTransactionProposal,
+    output: TransactionOutput,
+    // TODO: or call it vote?
+    signatures: BTreeMap<AccountAddress, ChannelTransactionSigs>,
+    lifecycle: ProposalLifecycle,
+}
+#[derive(Debug, Clone, Copy, Eq, PartialEq, PartialOrd, Serialize, Deserialize)]
+enum ProposalLifecycle {
+    Negotiating = 0,
+    Agreed,
+    Applying,
+    //    Fulfilling,
+    //    Fulfilled,
 }
 
 impl PendingTransaction {
-    pub fn add_signature(&mut self, sig: ChannelTransactionSigs) {
-        match self {
-            PendingTransaction::WaitForSig { signatures, .. } => {
-                signatures.insert(sig.address, sig);
-            }
-            PendingTransaction::WaitForApply { .. } => {
-                // TODO: debug_assert
-            }
-        }
-    }
-    pub fn get_signature(&self, address: &AccountAddress) -> Option<ChannelTransactionSigs> {
-        match self {
-            PendingTransaction::WaitForSig { signatures, .. } => signatures.get(address).cloned(),
-            PendingTransaction::WaitForApply { signatures, .. } => signatures.get(address).cloned(),
+    pub fn new(
+        proposal: ChannelTransactionProposal,
+        output: TransactionOutput,
+        signatures: BTreeMap<AccountAddress, ChannelTransactionSigs>,
+    ) -> Self {
+        Self {
+            proposal,
+            output,
+            signatures,
+            lifecycle: Negotiating,
         }
     }
 
-    pub fn fulfilled(&self) -> bool {
-        match self {
-            PendingTransaction::WaitForApply { .. } => true,
-            _ => false,
+    pub fn add_signature(&mut self, sig: ChannelTransactionSigs) {
+        if !self.signatures.contains_key(&sig.address) {
+            self.signatures.insert(sig.address, sig);
         }
     }
-    pub fn try_fulfill(&mut self, participants: &BTreeSet<AccountAddress>) -> bool {
-        match self {
-            PendingTransaction::WaitForSig {
-                signatures,
-                output,
-                proposal,
-            } => {
-                if signatures.len() == participants.len() {
-                    if participants
-                        .iter()
-                        .all(|addr| signatures.contains_key(addr))
-                    {
-                        *self = PendingTransaction::WaitForApply {
-                            proposal: proposal.clone(),
-                            signatures: signatures.clone(),
-                            output: output.clone(),
-                        };
-                    }
-                }
+    pub fn set_applying(&mut self) {
+        self.lifecycle = Applying;
+    }
+
+    pub fn get_signature(&self, address: &AccountAddress) -> Option<ChannelTransactionSigs> {
+        self.signatures.get(address).cloned()
+    }
+
+    pub fn newer_than(&self, other: &Self) -> bool {
+        self.lifecycle > other.lifecycle
+    }
+
+    pub fn is_negotiating(&self) -> bool {
+        self.lifecycle == Negotiating
+    }
+    pub fn consensus_reached(&self) -> bool {
+        self.lifecycle >= Agreed
+    }
+
+    pub fn try_reach_consensus(&mut self, participants: &BTreeSet<AccountAddress>) -> bool {
+        if self.lifecycle == Negotiating && self.signatures.len() == participants.len() {
+            if participants
+                .iter()
+                .all(|addr| self.signatures.contains_key(addr))
+            {
+                self.lifecycle = Agreed;
             }
-            PendingTransaction::WaitForApply { .. } => {
-                // TODO: debug_assert
-            }
-        };
-        self.fulfilled()
+        }
+        self.consensus_reached()
     }
 }
 
@@ -90,17 +93,12 @@ impl
         TransactionOutput,
         BTreeMap<AccountAddress, ChannelTransactionSigs>,
     ) {
-        match self {
-            PendingTransaction::WaitForSig {
-                proposal,
-                output,
-                signatures,
-            } => (proposal, output, signatures),
-            PendingTransaction::WaitForApply {
-                proposal,
-                output,
-                signatures,
-            } => (proposal, output, signatures),
-        }
+        let Self {
+            proposal,
+            output,
+            signatures,
+            ..
+        } = self;
+        (proposal, output, signatures)
     }
 }
