+++ sgwallet/Cargo.toml
@@ -7,6 +7,8 @@ publish = false
 edition = "2018"
 
 [dependencies]
+log = "0.4"
+uuid = "0.8"
 serde = {version = "1", default-features = false }
 hex = "0.3.2"
 rand = "0.6.5"
